requirements.txt
asgiref==3.8.1
Django==5.0.6
django-dotenv==1.4.2
django-taggit==5.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
